Library Research Skills e-learning course
Even though most students now use electronic resources to perform
research, there is still a need for students to understand how to find
information in a traditional library setting. This includes locating
books with a card catalog, finding them on the shelves, and locating
information in traditional hard-copy dictionaries and encyclopedias.